fixed language errors, fixed bug for places types
