+++ lib/example/showcase_app/screens/widgets/custom_bg.dart
